requirements.txt
python-telegram-bot==20.3
async
aioredis==2.0.1
hiredis==2.2.3
marshmallow==3.19.0
aiohttp-apispec==2.2.3
aiohttp==3.8.4